Major refactor to support correct timing
Most instructions take many cycles to complete and it's important that memory accesses occur on the correct particular cycles.

To date, Tetromino has executed the whole instruction atomically at the end whatever N number of cycles was right. This commit introduces a new architecture that treats each instruction as a series of steps, each of which takes exactly one machine cycle. Most instructions still don't actually implement the correct timing, but the foundations are in place.

+++ pkg/gb/cpu/instruction_metadata.go
@@ -11,7 +11,7 @@ import (
 
 type metadata struct {
 	Mnemonic         string   `json:"mnemonic"`
-	Length           uint16   `json:"length"`
+	Length           int      `json:"length"`
 	Cycles           []int    `json:"cycles"`
 	Flags            []string `json:"flags"`
 	Addr             string   `json:"addr"`
